--- src/main/java/com/noint/shelterzoo/user/dto/UserDTO.java
@@ -1,12 +1,0 @@
-package com.noint.shelterzoo.user.dto;
-
-import lombok.Data;
-
-public class UserDTO {
-    @Data
-    public static class Signup {
-        private String email;
-        private String nickname;
-        private String password;
-    }
-}
